# server/config/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "LMS Application API",
      "description": "LMS Application API",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:3000",
    "basePath": "/api/",
...
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "securityDefinitions": {
      "AuthHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "paths": {
      "/user": {
        "get": {
          "tags": ["User"],
          "description": "Retrieve All The Users",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "List of All The Users",
              "schema": {
                "type": "Array",
                "$ref": "#/definitions/User"
              }
            },
            "404": {
              "description": "Data Not Found!!"
            },
            "401": {
              "description": "Unauthorized!!"
            },
            "500": {
              "description": "Something Went Wrong!!"
            }
          }
        }
      },
      "/user/{id}": {
        "get": {
          "tags": ["User"],
          "description": "Retrieve Single User",
          "produces": [
            "application/json"
          ],
          "parameters": [{
            "name": "id",
            "in": "path",
            "description": "id of user",
            "schema": {
              "type": "integer",
              "required": "true"
            }
          }],
          "responses": {
            "200": {
              "description": "Succesfully Fetched Single User",
              "schema": {
                "type": "Array",
                "$ref": "#/definitions/User"
              }
            },
            "404": {
              "description": "Data Not Found!!"
            },
            "401": {
              "description": "Unauthorized!!"
            },
            "500": {
              "description": "Something Went Wrong!!"
            }
          }
        },
        "put": {
          "tags": ["User"],
          "description": "Update Single User by Admin",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "id of user",
              "schema": {
                "type": "integer",
                "required": "true"
              }
            },
            {
              "name": "body",
              "in": "body",
              "description": "body for update",
              "schema": {
                "type": "object",
                "required": "true",
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "User is updated",
              "schema": {
                "type": "object",
                "$ref": "#/definitions/User"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Validation Error || Something Went Wrong"
            }
          }
        },
        "delete": {
          "tags": ["User"],
          "description": "Delete Single User",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "id of user",
              "schema": {
                "type": "integer",
                "required": "true"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "User is Deleted"
            },
            "401": {
              "description": "Unauthorized"
            },
            "400": {
              "description": "Not Found"
            },
            "500": {
              "description": "Something Went Wrong"
            }
          }
        }
      },      
...
    },
    "definitions": {
      "User": {
        "required": [
          "email",
          "first_name",
          "last_name",
          "mobile",
          "role"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "uniqueItems": true
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "mobile": {
            "type": "string",
            "uniqueItems": true
          },
          "email": {
            "type": "string",
            "uniqueItems": true
          },
          "profile_img": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "speciality": {
            "type": "string"
          },
          "additional_qualifications": {
            "type": "string"
          },
          "rcs_publication_details": {
            "type": "string"
          },
          "med_council_reg_no": {
            "type": "string"
          },
          "hospital_details": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "is_admin": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        }
      },
...
  }
}